+++ apps/blogit/utils/updateMessage.ts
@@ -4,7 +4,7 @@ import slackClient from "./slackClient";
 export default async (
   payload: ChatPostMessageArguments["blocks"],
   channel: string,
-  message_ts: string
+  message_ts: string,
 ) => {
   return await slackClient.chat.update({
     text: "Hey it's me, Blogit!",