+++ apps/docs/app/page.tsx
@@ -1,10 +1,10 @@
-import { ComingSoonButton, Header } from "ui";
+import { ComingSoonButton, Header } from "@rubriclab/ui";
 
 export default function Page() {
   return (
-    <>
+    <div className="flex items-center flex-col justify-center">
       <Header text="Docs" />
       <ComingSoonButton />
-    </>
+    </div>
   );
 }