+++ packages/@rubriclab/ui/tailwind.config.js
@@ -1,0 +1,8 @@
+const sharedConfig = require("@rubriclab/tailwind-config");
+
+module.exports = {
+  presets: [sharedConfig],
+  theme: {
+    extend: {},
+  },
+};