+++ packages/create-rubric-app/templates/fullstack/constants/fonts.ts
@@ -1,0 +1,4 @@
+export const FONTS = {
+  // This can be replaced with local "/fonts/CalSans-SemiBold.ttf" once you re-download the font as explained in app/layout.tsx
+  calSansURL: "https://rubriclab.com/fonts/CalSans-SemiBold.ttf",
+};