+++ packages/create-rubric-app/templates/fullstack/postcss.config.js
@@ -1,1 +1,1 @@
-module.exports = require("tailwind-config/postcss.config");
+module.exports = require("@rubriclab/tailwind-config/postcss.cjs");