+++ packages/ui/index.tsx
@@ -1,6 +1,0 @@
-import * as React from "react";
-
-// component exports
-export * from "./ContactButton";
-export * from "./ComingSoonButton";
-export * from "./Header";