+++ packages/ui/tailwind.config.js
@@ -1,8 +1,0 @@
-const sharedConfig = require("tailwind-config/tailwind.config");
-
-module.exports = {
-  presets: [sharedConfig],
-  theme: {
-    extend: {},
-  },
-};