+++ packages/create-rubric-app/CHANGELOG.md
@@ -1,5 +1,113 @@
 # create-rubric-app
 
+## 0.1.7
+
+### Patch Changes
+
+- comment font bug
+
+## 0.1.6
+
+### Patch Changes
+
+- add font back
+
+## 0.1.5
+
+### Patch Changes
+
+- remove font
+
+## 0.1.4
+
+### Patch Changes
+
+- change package.json rubric deps to latest
+
+## 0.1.3
+
+### Patch Changes
+
+- font is buggy
+
+## 0.1.2
+
+### Patch Changes
+
+- updated styling
+
+## 0.1.1
+
+### Patch Changes
+
+- fix font loading
+
+## 0.1.0
+
+### Minor Changes
+
+- minor release all packages
+
+## 0.0.18
+
+### Patch Changes
+
+- upgraded boilerplate
+
+## 0.0.17
+
+### Patch Changes
+
+- transpile ui
+
+## 0.0.16
+
+### Patch Changes
+
+- update template
+
+## 0.0.15
+
+### Patch Changes
+
+- rm node modules
+
+## 0.0.14
+
+### Patch Changes
+
+- rm yarn lock
+
+## 0.0.13
+
+### Patch Changes
+
+- update includes in tsconfig
+
+## 0.0.12
+
+### Patch Changes
+
+- bump configs and cli
+
+## 0.0.11
+
+### Patch Changes
+
+- cjs
+
+## 0.0.10
+
+### Patch Changes
+
+- bump configs
+
+## 0.0.9
+
+### Patch Changes
+
+- remove @rubriclab/ prefix in favour of simply npx create-rubric-app
+
 ## 0.0.8
 
 ### Patch Changes